pyproject.toml
[project]
name = "gmail-quickstart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "google-api-python-client>=2.188.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-json-logger>=4.0.0",
]
